Remove Blue specific code
