+++ .gitignore
@@ -6,4 +6,3 @@ src/glyphs.c
 src/glyphs.h
 .idea
 nano-sdk
-blue-sdk
