+++ src/stellar_format.h
@@ -32,6 +32,7 @@ typedef void (*format_function_t)(tx_context_t *txCtx);
 extern format_function_t formatter;
 extern format_function_t formatter_stack[MAX_FORMATTERS_PER_OPERATION];
 extern int8_t formatter_index;
+extern uint8_t current_data_index;
 
 /* the current details printed by the formatter */
 extern char opCaption[OPERATION_CAPTION_MAX_SIZE];
@@ -41,6 +42,8 @@ extern char detailValue[DETAIL_VALUE_MAX_SIZE];
 void format_confirm_transaction(tx_context_t *txCtx);
 void format_confirm_operation(tx_context_t *txCtx);
 void format_confirm_transaction_details(tx_context_t *txCtx);
+void set_state_data(bool forward);
+void ui_approve_tx_next_screen(tx_context_t *txCtx);
 
 void format_confirm_hash_warning(tx_context_t *txCtx);
 #endif
